Extracted abstract class Level

+++ src/Paddle.java
@@ -7,7 +7,7 @@ public class Paddle extends GCompound {
     private GImage body;
 
     private int width,height;
-    public Paddle(int width, int height) {
+    public Paddle(double width, double height) {
         body = new GImage(PADDLE_IMAGE);
         this.body.setSize(width, height);
         add(this.body);
